requirements.txt
# Gmail API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
google-auth==2.45.0

# AI Providers
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
anthropic==0.39.0

# Data Processing
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-core==2.14.1

# Analysis & Visualization (for precision-recall optimization)
pandas==2.3.3
matplotlib==3.10.8
scikit-learn==1.8.0
numpy==1.26.4
scipy==1.16.3

# Utilities
tqdm==4.66.1
colorama==0.4.6
